+++ examples/single_sensor/single_sensor.ino
@@ -39,9 +39,9 @@
 // The name of this program file
 const char* sketchName = "single_sensor.ino";
 
-const long   serialBaud = 115200;  // Baud rate for debugging
-const int8_t greenLED   = 8;       // Pin for the green LED
-const int8_t redLED     = 9;       // Pin for the red LED
+const int32_t serialBaud = 115200;  // Baud rate for debugging
+const int8_t  greenLED   = 8;       // Pin for the green LED
+const int8_t  redLED     = 9;       // Pin for the red LED
 /** End [sketch_info] */
